+++ lib/omq/routing/router.rb
@@ -15,8 +15,8 @@ module OMQ
       #
       def initialize(engine)
         @engine                  = engine
-        @recv_queue              = Async::LimitedQueue.new(engine.options.recv_hwm)
-        @send_queue              = Async::LimitedQueue.new(engine.options.send_hwm)
+        @recv_queue              = Routing.build_queue(engine.options.recv_hwm, :block)
+        @send_queue              = Routing.build_queue(engine.options.send_hwm, :block)
         @connections_by_identity = {}
         @identity_by_connection  = {}
         @tasks                   = []