+++ lib/omq/routing/xsub.rb
@@ -14,8 +14,8 @@ module OMQ
       def initialize(engine)
         @engine            = engine
         @connections       = []
-        @recv_queue        = Async::LimitedQueue.new(engine.options.recv_hwm)
-        @send_queue        = Async::LimitedQueue.new(engine.options.send_hwm)
+        @recv_queue        = Routing.build_queue(engine.options.recv_hwm, engine.options.on_mute)
+        @send_queue        = Routing.build_queue(engine.options.send_hwm, :block)
         @tasks             = []
         @send_pump_started = false
         @send_pump_idle    = true